services/ingestion/requirements.txt
kafka-python
watchdog
python-dotenv
